Organize xml module files
